requirements.txt
Flask==2.2.3
google-cloud-vision==3.4.0
openai==0.27.2
requests==2.28.2